move BitcoinTx out of bridge-primitives
