+++ crates/bridge-proof/protocol2/src/lib.rs
@@ -12,6 +12,7 @@ mod error;
 mod prover;
 mod statement;
 mod test_data;
+mod tx;
 mod tx_inclusion_proof;
 mod tx_info;
