+++ crates/bridge-proof/protocol2/src/tx_inclusion_proof.rs
@@ -1,6 +1,5 @@
 use bitcoin::{block::Header, hashes::Hash, Transaction};
 use borsh::{BorshDeserialize, BorshSerialize};
-use strata_bridge_primitives::bitcoin::BitcoinTx;
 use strata_primitives::{
     buf::Buf32,
     hash::sha256d,
@@ -8,6 +7,8 @@ use strata_primitives::{
 };
 use strata_proofimpl_btc_blockspace::block::witness_commitment_from_coinbase;
 use strata_state::l1::{L1TxInclusionProof, L1TxProof, L1WtxProof};
+
+use crate::tx::BitcoinTx;
 
 /// A transaction along with its [L1TxInclusionProof], parameterized by a `Marker` type
 /// (either [`TxIdMarker`] or [`WtxIdMarker`]).
