+++ crates/primitives/Cargo.toml
@@ -17,13 +17,9 @@ bitcoin = { workspace = true, features = ["rand-std"] }
 bitcoin-bosd.workspace = true
 bitcoin-script.workspace = true
 bitvm.workspace = true
-borsh.workspace = true
 musig2 = { workspace = true, features = ["serde"] }
 rkyv.workspace = true
 secp256k1 = { workspace = true, features = ["global-context", "rand-std"] }
 serde.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
-
-[dev-dependencies]
-strata-bridge-test-utils.workspace = true
